requirements.txt
yt-dlp
openai-whisper
openai
faiss-cpu
sentence-transformers
langchain
gtts
numpy
json